feat(kag-solver): query schema接口和graph write接口重写
