+++ kag/examples/TwoWikiTest/builder/ckpt/kag_checkpoint_0_1.ckpt
@@ -1,8 +1,8 @@
-{"id": "d99687bd5e66d7d9d86ec9232e599a453f8cf4c790170429b189808516099e21", "value": {"abstract": {"id": "“00后”女兵逆袭体", "name": "“00后”女兵逆袭体", "content": "URL: http:"}, "graph_stat": {"num_nodes": 10, "num_edges": 16, "num_subgraphs": 1}}}
-{"id": "1e42a763a1a22f6096463ce89c0753f0325399783a1402df559a2e94d8bc9f3c", "value": {"abstract": {"id": "新华时评丨行全球惯例", "name": "新华时评丨行全球惯例", "content": "URL: http:"}, "graph_stat": {"num_nodes": 31, "num_edges": 50, "num_subgraphs": 1}}}
-{"id": "2e7189c7784ee89a28af62543553472e8725c8acb4ddb175ac32a8dd8e93b55f", "value": {"abstract": {"id": "在山巅“走向深蓝”—", "name": "在山巅“走向深蓝”—", "content": "URL: http:"}, "graph_stat": {"num_nodes": 38, "num_edges": 65, "num_subgraphs": 1}}}
-{"id": "a35a32c07c8497aae2e42b7eb933d2b18c878a19f5b538a37ce25f4f7eb678ed", "value": {"abstract": {"id": "2025年6月9日外", "name": "2025年6月9日外", "content": "URL: http:"}, "graph_stat": {"num_nodes": 61, "num_edges": 96, "num_subgraphs": 1}}}
-{"id": "09d7df16d246abd4c1bccf89127894c69d558c81fae9ada9eca2c1a1a3b19753", "value": {"abstract": {"id": "新疆哈密军分区建立“", "name": "新疆哈密军分区建立“", "content": "URL: http:"}, "graph_stat": {"num_nodes": 37, "num_edges": 59, "num_subgraphs": 1}}}
-{"id": "9c510fbbf8d1bd1295ae59bbf2dbca3c0b1e822283193ed857304663f251d8e8", "value": {"abstract": {"id": "F-35战斗机的“天", "name": "F-35战斗机的“天", "content": "URL: http:"}, "graph_stat": {"num_nodes": 45, "num_edges": 79, "num_subgraphs": 1}}}
-{"id": "5f36ef5cb798670db000341a8022db734dc18eb0200be13ed83aae94818e5863", "value": {"abstract": {"id": "新加坡海军现代化建设", "name": "新加坡海军现代化建设", "content": "URL: http:"}, "graph_stat": {"num_nodes": 56, "num_edges": 97, "num_subgraphs": 1}}}
-{"id": "58e46b6fb65457c93007245d84a6f447a1b5f668e0e041da92a383c6c33d627d", "value": {"abstract": {"id": "新华时评丨让文明对话", "name": "新华时评丨让文明对话", "content": "URL: http:"}, "graph_stat": {"num_nodes": 43, "num_edges": 72, "num_subgraphs": 1}}}
+{"id": "d99687bd5e66d7d9d86ec9232e599a453f8cf4c790170429b189808516099e21", "value": {"abstract": {"id": "“00后”女兵逆袭体", "name": "“00后”女兵逆袭体", "content": "URL: http:"}, "graph_stat": {"num_nodes": 16, "num_edges": 23, "num_subgraphs": 1}}}
+{"id": "1e42a763a1a22f6096463ce89c0753f0325399783a1402df559a2e94d8bc9f3c", "value": {"abstract": {"id": "新华时评丨行全球惯例", "name": "新华时评丨行全球惯例", "content": "URL: http:"}, "graph_stat": {"num_nodes": 27, "num_edges": 41, "num_subgraphs": 1}}}
+{"id": "9c510fbbf8d1bd1295ae59bbf2dbca3c0b1e822283193ed857304663f251d8e8", "value": {"abstract": {"id": "F-35战斗机的“天", "name": "F-35战斗机的“天", "content": "URL: http:"}, "graph_stat": {"num_nodes": 46, "num_edges": 79, "num_subgraphs": 1}}}
+{"id": "2e7189c7784ee89a28af62543553472e8725c8acb4ddb175ac32a8dd8e93b55f", "value": {"abstract": {"id": "在山巅“走向深蓝”—", "name": "在山巅“走向深蓝”—", "content": "URL: http:"}, "graph_stat": {"num_nodes": 42, "num_edges": 64, "num_subgraphs": 1}}}
+{"id": "09d7df16d246abd4c1bccf89127894c69d558c81fae9ada9eca2c1a1a3b19753", "value": {"abstract": {"id": "新疆哈密军分区建立“", "name": "新疆哈密军分区建立“", "content": "URL: http:"}, "graph_stat": {"num_nodes": 35, "num_edges": 55, "num_subgraphs": 1}}}
+{"id": "a35a32c07c8497aae2e42b7eb933d2b18c878a19f5b538a37ce25f4f7eb678ed", "value": {"abstract": {"id": "2025年6月9日外", "name": "2025年6月9日外", "content": "URL: http:"}, "graph_stat": {"num_nodes": 30, "num_edges": 44, "num_subgraphs": 1}}}
+{"id": "58e46b6fb65457c93007245d84a6f447a1b5f668e0e041da92a383c6c33d627d", "value": {"abstract": {"id": "新华时评丨让文明对话", "name": "新华时评丨让文明对话", "content": "URL: http:"}, "graph_stat": {"num_nodes": 66, "num_edges": 112, "num_subgraphs": 1}}}
+{"id": "5f36ef5cb798670db000341a8022db734dc18eb0200be13ed83aae94818e5863", "value": {"abstract": {"id": "新加坡海军现代化建设", "name": "新加坡海军现代化建设", "content": "URL: http:"}, "graph_stat": {"num_nodes": 66, "num_edges": 116, "num_subgraphs": 1}}}
