+++ kag/examples/TwoWikiTest/builder/ckpt/splitter/kag_checkpoint_0_1.ckpt.lock
@@ -1,1 +1,1 @@
- 417126
+ 510299
