+++ kag/examples/TwoWikiTest/builder/ckpt/writer/kag_checkpoint_0_1.ckpt.lock
@@ -1,1 +1,1 @@
- 417126
+ 510299
