--- kag/jiuyuansolver/schema3.py
@@ -1,1 +1,0 @@
-{"spgTypes":[{"@type":"TEXT","basicType":"TEXT","basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本"},"spgTypeEnum":"BASIC_TYPE","properties":[],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{"constrains":[{"id":"REQUIRE","name":"Required","nameZh":"值非空"},{"id":"UNIQUE","name":"Unique","nameZh":"值唯一"},{"id":"ENUM","name":"Enum","nameZh":"枚举"},{"id":"MULTIVALUE","name":"Multi value","nameZh":"多值"},{"id":"REGULAR","name":"Regular match","nameZh":"正则匹配"}]}},{"@type":"LONG","basicType":"LONG","basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Integer","identityType":"SPG_TYPE"},"nameZh":"整型"},"spgTypeEnum":"BASIC_TYPE","properties":[],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":4,"alterId":4},"extInfo":{"constrains":[{"id":"REQUIRE","name":"Required","nameZh":"值非空"},{"id":"ENUM","name":"Enum","nameZh":"枚举"},{"id":"MINIMUM_GT","name":"Greater than","nameZh":"大于"},{"id":"MINIMUM_GT_OE","name":"Greater than or equal","nameZh":"大于等于"},{"id":"MAXIMUM_LT","name":"Less than","nameZh":"小于"},{"id":"MAXIMUM_LT_OE","name":"Less than or equal","nameZh":"小于等于"}]}},{"@type":"DOUBLE","basicType":"DOUBLE","basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Float","identityType":"SPG_TYPE"},"nameZh":"浮点数"},"spgTypeEnum":"BASIC_TYPE","properties":[],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":5,"alterId":5},"extInfo":{"constrains":[{"id":"REQUIRE","name":"Required","nameZh":"值非空"},{"id":"ENUM","name":"Enum","nameZh":"枚举"},{"id":"MINIMUM_GT","name":"Greater than","nameZh":"大于"},{"id":"MINIMUM_GT_OE","name":"Greater than or equal","nameZh":"大于等于"},{"id":"MAXIMUM_LT","name":"Less than","nameZh":"小于"},{"id":"MAXIMUM_LT_OE","name":"Less than or equal","nameZh":"小于等于"}]}},{"@type":"STANDARD_TYPE","spreadable":false,"constraintItems":[{"@type":"REGULAR","regularPattern":"[1,2][0-9][0-9][0-9](0[1-9]|1[0-2])(0[1-9]|[1,2][0-9]|3[0,1])","pattern":"[1,2][0-9][0-9][0-9](0[1-9]|1[0-2])(0[1-9]|[1,2][0-9]|3[0,1])"}],"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期","desc":"8位数字组成的日期"},"parentTypeInfo":{"uniqueId":19,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,19]},"spgTypeEnum":"STANDARD_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期","desc":"8位数字组成的日期"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":19,"alterId":19}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期","desc":"8位数字组成的日期"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":19,"alterId":19}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期","desc":"8位数字组成的日期"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":19,"alterId":19}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}}],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":19,"alterId":19},"extInfo":{"constrains":[{"id":"REGULAR","name":"Regular match","nameZh":"正则匹配","value":"[1,2][0-9][0-9][0-9](0[1-9]|1[0-2])(0[1-9]|[1,2][0-9]|3[0,1])"}]}},{"@type":"STANDARD_TYPE","spreadable":true,"constraintItems":[{"@type":"MULTI_VALUE"},{"@type":"REGULAR","regularPattern":"^((13[0-9])|(14[5,7,9])|(15([0-3]|[5-9]))|(16[5,6])|(17[0-8])|(18[0-9])|(19[1,5,8,9]))[0-9]{8}$","pattern":"^((13[0-9])|(14[5,7,9])|(15([0-3]|[5-9]))|(16[5,6])|(17[0-8])|(18[0-9])|(19[1,5,8,9]))[0-9]{8}$"}],"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"ChinaMobile","identityType":"SPG_TYPE"},"nameZh":"国内手机号","desc":"中国国内使用的手机号码由11位数字组成"},"parentTypeInfo":{"uniqueId":10,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,10]},"spgTypeEnum":"STANDARD_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"ChinaMobile","identityType":"SPG_TYPE"},"nameZh":"国内手机号","desc":"中国国内使用的手机号码由11位数字组成"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":10,"alterId":10}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"ChinaMobile","identityType":"SPG_TYPE"},"nameZh":"国内手机号","desc":"中国国内使用的手机号码由11位数字组成"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":10,"alterId":10}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"ChinaMobile","identityType":"SPG_TYPE"},"nameZh":"国内手机号","desc":"中国国内使用的手机号码由11位数字组成"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":10,"alterId":10}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}}],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":10,"alterId":10},"extInfo":{"spreadable":true,"constrains":[{"id":"MULTIVALUE","name":"Multi value","nameZh":"多值"},{"id":"REGULAR","name":"Regular match","nameZh":"正则匹配","value":"^((13[0-9])|(14[5,7,9])|(15([0-3]|[5-9]))|(16[5,6])|(17[0-8])|(18[0-9])|(19[1,5,8,9]))[0-9]{8}$"}]}},{"@type":"STANDARD_TYPE","spreadable":true,"constraintItems":[{"@type":"REGULAR","regularPattern":"^(400[0-9]{7})|(800[0-9]{7})|(0[0-9]{2,3}-[0-9]{7,8})|((13[0-9])|(14[5,7,9])|(15([0-3]|[5-9]))|(16[5,6])|(17[0-8])|(18[0-9])|(19[1,5,8,9]))[0-9]{8}$","pattern":"^(400[0-9]{7})|(800[0-9]{7})|(0[0-9]{2,3}-[0-9]{7,8})|((13[0-9])|(14[5,7,9])|(15([0-3]|[5-9]))|(16[5,6])|(17[0-8])|(18[0-9])|(19[1,5,8,9]))[0-9]{8}$"}],"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"ChinaTelCode","identityType":"SPG_TYPE"},"nameZh":"国内通讯号","desc":"国内通讯号码包含常见座机和手机号码"},"parentTypeInfo":{"uniqueId":27,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,27]},"spgTypeEnum":"STANDARD_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"ChinaTelCode","identityType":"SPG_TYPE"},"nameZh":"国内通讯号","desc":"国内通讯号码包含常见座机和手机号码"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":27,"alterId":27}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"ChinaTelCode","identityType":"SPG_TYPE"},"nameZh":"国内通讯号","desc":"国内通讯号码包含常见座机和手机号码"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":27,"alterId":27}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"ChinaTelCode","identityType":"SPG_TYPE"},"nameZh":"国内通讯号","desc":"国内通讯号码包含常见座机和手机号码"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":27,"alterId":27}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}}],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":27,"alterId":27},"extInfo":{"spreadable":true,"constrains":[{"id":"REGULAR","name":"Regular match","nameZh":"正则匹配","value":"^(400[0-9]{7})|(800[0-9]{7})|(0[0-9]{2,3}-[0-9]{7,8})|((13[0-9])|(14[5,7,9])|(15([0-3]|[5-9]))|(16[5,6])|(17[0-8])|(18[0-9])|(19[1,5,8,9]))[0-9]{8}$"}]}},{"@type":"STANDARD_TYPE","spreadable":true,"constraintItems":[{"@type":"MULTI_VALUE"},{"@type":"REGULAR","regularPattern":"^([a-zA-Z0-9]*[-_.]?[a-zA-Z0-9]+)*@([a-zA-Z0-9]*[-_]?[a-zA-Z0-9]+)+[.][A-Za-z]{2,3}([.][A-Za-z]{2})?$","pattern":"^([a-zA-Z0-9]*[-_.]?[a-zA-Z0-9]+)*@([a-zA-Z0-9]*[-_]?[a-zA-Z0-9]+)+[.][A-Za-z]{2,3}([.][A-Za-z]{2})?$"}],"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Email","identityType":"SPG_TYPE"},"nameZh":"电子邮箱","desc":"电子邮箱地址"},"parentTypeInfo":{"uniqueId":11,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,11]},"spgTypeEnum":"STANDARD_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Email","identityType":"SPG_TYPE"},"nameZh":"电子邮箱","desc":"电子邮箱地址"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":11,"alterId":11}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Email","identityType":"SPG_TYPE"},"nameZh":"电子邮箱","desc":"电子邮箱地址"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":11,"alterId":11}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Email","identityType":"SPG_TYPE"},"nameZh":"电子邮箱","desc":"电子邮箱地址"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":11,"alterId":11}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}}],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":11,"alterId":11},"extInfo":{"spreadable":true,"constrains":[{"id":"MULTIVALUE","name":"Multi value","nameZh":"多值"},{"id":"REGULAR","name":"Regular match","nameZh":"正则匹配","value":"^([a-zA-Z0-9]*[-_.]?[a-zA-Z0-9]+)*@([a-zA-Z0-9]*[-_]?[a-zA-Z0-9]+)+[.][A-Za-z]{2,3}([.][A-Za-z]{2})?$"}]}},{"@type":"STANDARD_TYPE","spreadable":false,"constraintItems":[{"@type":"REGULAR","regularPattern":"^([0-9]{10})|([0-9]{13})$","pattern":"^([0-9]{10})|([0-9]{13})$"}],"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Timestamp","identityType":"SPG_TYPE"},"nameZh":"时间戳","desc":"10位或者13位的时间戳"},"parentTypeInfo":{"uniqueId":29,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,29]},"spgTypeEnum":"STANDARD_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Timestamp","identityType":"SPG_TYPE"},"nameZh":"时间戳","desc":"10位或者13位的时间戳"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":29,"alterId":29}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Timestamp","identityType":"SPG_TYPE"},"nameZh":"时间戳","desc":"10位或者13位的时间戳"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":29,"alterId":29}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Timestamp","identityType":"SPG_TYPE"},"nameZh":"时间戳","desc":"10位或者13位的时间戳"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":29,"alterId":29}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}}],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":29,"alterId":29},"extInfo":{"constrains":[{"id":"REGULAR","name":"Regular match","nameZh":"正则匹配","value":"^([0-9]{10})|([0-9]{13})$"}]}},{"@type":"STANDARD_TYPE","spreadable":true,"constraintItems":[{"@type":"MULTI_VALUE"},{"@type":"REGULAR","regularPattern":"^[1-9]{1}[0-9]{5}(19|20)[0-9]{2}((0[1-9]{1})|(1[0-2]{1}))((0[1-9]{1})|([1-2]{1}[0-9]{1}|(3[0-1]{1})))[0-9]{3}[0-9xX]{1}$","pattern":"^[1-9]{1}[0-9]{5}(19|20)[0-9]{2}((0[1-9]{1})|(1[0-2]{1}))((0[1-9]{1})|([1-2]{1}[0-9]{1}|(3[0-1]{1})))[0-9]{3}[0-9xX]{1}$"}],"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"IdCardNo","identityType":"SPG_TYPE"},"nameZh":"身份证","desc":"中国身份证号码一般由18位数字和字母组成"},"parentTypeInfo":{"uniqueId":13,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,13]},"spgTypeEnum":"STANDARD_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"IdCardNo","identityType":"SPG_TYPE"},"nameZh":"身份证","desc":"中国身份证号码一般由18位数字和字母组成"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":13,"alterId":13}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"IdCardNo","identityType":"SPG_TYPE"},"nameZh":"身份证","desc":"中国身份证号码一般由18位数字和字母组成"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":13,"alterId":13}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"IdCardNo","identityType":"SPG_TYPE"},"nameZh":"身份证","desc":"中国身份证号码一般由18位数字和字母组成"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":13,"alterId":13}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}}],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":13,"alterId":13},"extInfo":{"spreadable":true,"constrains":[{"id":"MULTIVALUE","name":"Multi value","nameZh":"多值"},{"id":"REGULAR","name":"Regular match","nameZh":"正则匹配","value":"^[1-9]{1}[0-9]{5}(19|20)[0-9]{2}((0[1-9]{1})|(1[0-2]{1}))((0[1-9]{1})|([1-2]{1}[0-9]{1}|(3[0-1]{1})))[0-9]{3}[0-9xX]{1}$"}]}},{"@type":"STANDARD_TYPE","spreadable":true,"constraintItems":[{"@type":"MULTI_VALUE"},{"@type":"REGULAR","regularPattern":"([A-Fa-f0-9]{2}-){5}[A-Fa-f0-9]{2}","pattern":"([A-Fa-f0-9]{2}-){5}[A-Fa-f0-9]{2}"}],"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"MacAddress","identityType":"SPG_TYPE"},"nameZh":"MAC地址","desc":"网卡MAC地址"},"parentTypeInfo":{"uniqueId":14,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,14]},"spgTypeEnum":"STANDARD_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"MacAddress","identityType":"SPG_TYPE"},"nameZh":"MAC地址","desc":"网卡MAC地址"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":14,"alterId":14}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"MacAddress","identityType":"SPG_TYPE"},"nameZh":"MAC地址","desc":"网卡MAC地址"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":14,"alterId":14}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"MacAddress","identityType":"SPG_TYPE"},"nameZh":"MAC地址","desc":"网卡MAC地址"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":14,"alterId":14}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}}],"relations":[],"advancedConfig":{"visibleScope":"PUBLIC"},"ontologyId":{"uniqueId":14,"alterId":14},"extInfo":{"spreadable":true,"constrains":[{"id":"MULTIVALUE","name":"Multi value","nameZh":"多值"},{"id":"REGULAR","name":"Regular match","nameZh":"正则匹配","value":"([A-Fa-f0-9]{2}-){5}[A-Fa-f0-9]{2}"}]}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Chunk","identityType":"SPG_TYPE"},"nameZh":"文本块"},"parentTypeInfo":{"uniqueId":2025062280074614,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062280074614]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Chunk","identityType":"SPG_TYPE"},"nameZh":"文本块"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062280074614,"alterId":2025062280074614}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Chunk","identityType":"SPG_TYPE"},"nameZh":"文本块"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062280074614,"alterId":2025062280074614}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Chunk","identityType":"SPG_TYPE"},"nameZh":"文本块"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062280074614,"alterId":2025062280074614}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Chunk","identityType":"SPG_TYPE"},"nameZh":"文本块"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062280074614,"alterId":2025062280074614}},"basicInfo":{"name":{"@type":"PREDICATE","name":"content","identityType":"PREDICATE"},"nameZh":"内容"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062256354342,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062257391990,"alterId":2025062257391990},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062280074614,"alterId":2025062280074614},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"NaturalScience","identityType":"SPG_TYPE"},"nameZh":"自然科学"},"parentTypeInfo":{"uniqueId":2025062230155913,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062230155913]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"NaturalScience","identityType":"SPG_TYPE"},"nameZh":"自然科学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062230155913,"alterId":2025062230155913}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"NaturalScience","identityType":"SPG_TYPE"},"nameZh":"自然科学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062230155913,"alterId":2025062230155913}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"NaturalScience","identityType":"SPG_TYPE"},"nameZh":"自然科学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062230155913,"alterId":2025062230155913}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"NaturalScience","identityType":"SPG_TYPE"},"nameZh":"自然科学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062230155913,"alterId":2025062230155913}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062213488874,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062215257489,"alterId":2025062215257489},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"NaturalScience","identityType":"SPG_TYPE"},"nameZh":"自然科学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062230155913,"alterId":2025062230155913}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062225219531,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062215584257,"alterId":2025062215584257},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062230155913,"alterId":2025062230155913},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"SemanticConcept","identityType":"SPG_TYPE"},"nameZh":"语义概念"},"parentTypeInfo":{"uniqueId":2025062269362846,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062269362846]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"SemanticConcept","identityType":"SPG_TYPE"},"nameZh":"语义概念"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062269362846,"alterId":2025062269362846}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"SemanticConcept","identityType":"SPG_TYPE"},"nameZh":"语义概念"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062269362846,"alterId":2025062269362846}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"SemanticConcept","identityType":"SPG_TYPE"},"nameZh":"语义概念"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062269362846,"alterId":2025062269362846}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"SemanticConcept","identityType":"SPG_TYPE"},"nameZh":"语义概念"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062269362846,"alterId":2025062269362846}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"内容"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062255578209,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062215680151,"alterId":2025062215680151},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062269362846,"alterId":2025062269362846},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Works","identityType":"SPG_TYPE"},"nameZh":"作品"},"parentTypeInfo":{"uniqueId":2025062247130955,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062247130955]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Works","identityType":"SPG_TYPE"},"nameZh":"作品"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247130955,"alterId":2025062247130955}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Works","identityType":"SPG_TYPE"},"nameZh":"作品"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247130955,"alterId":2025062247130955}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Works","identityType":"SPG_TYPE"},"nameZh":"作品"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247130955,"alterId":2025062247130955}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Works","identityType":"SPG_TYPE"},"nameZh":"作品"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247130955,"alterId":2025062247130955}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062211685486,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062237311127,"alterId":2025062237311127},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Works","identityType":"SPG_TYPE"},"nameZh":"作品"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247130955,"alterId":2025062247130955}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062298283433,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062289167670,"alterId":2025062289167670},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062247130955,"alterId":2025062247130955},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Creature","identityType":"SPG_TYPE"},"nameZh":"生物"},"parentTypeInfo":{"uniqueId":2025062227789467,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062227789467]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Creature","identityType":"SPG_TYPE"},"nameZh":"生物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062227789467,"alterId":2025062227789467}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Creature","identityType":"SPG_TYPE"},"nameZh":"生物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062227789467,"alterId":2025062227789467}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Creature","identityType":"SPG_TYPE"},"nameZh":"生物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062227789467,"alterId":2025062227789467}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Creature","identityType":"SPG_TYPE"},"nameZh":"生物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062227789467,"alterId":2025062227789467}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062265503529,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062214220025,"alterId":2025062214220025},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Creature","identityType":"SPG_TYPE"},"nameZh":"生物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062227789467,"alterId":2025062227789467}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062238888408,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062293058129,"alterId":2025062293058129},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062227789467,"alterId":2025062227789467},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Building","identityType":"SPG_TYPE"},"nameZh":"建筑"},"parentTypeInfo":{"uniqueId":2025062274417542,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062274417542]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Building","identityType":"SPG_TYPE"},"nameZh":"建筑"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062274417542,"alterId":2025062274417542}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Building","identityType":"SPG_TYPE"},"nameZh":"建筑"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062274417542,"alterId":2025062274417542}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Building","identityType":"SPG_TYPE"},"nameZh":"建筑"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062274417542,"alterId":2025062274417542}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Building","identityType":"SPG_TYPE"},"nameZh":"建筑"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062274417542,"alterId":2025062274417542}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062287116714,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062254004552,"alterId":2025062254004552},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Building","identityType":"SPG_TYPE"},"nameZh":"建筑"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062274417542,"alterId":2025062274417542}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062211336216,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062272936947,"alterId":2025062272936947},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062274417542,"alterId":2025062274417542},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Medicine","identityType":"SPG_TYPE"},"nameZh":"药物"},"parentTypeInfo":{"uniqueId":2025062221855140,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062221855140]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Medicine","identityType":"SPG_TYPE"},"nameZh":"药物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062221855140,"alterId":2025062221855140}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Medicine","identityType":"SPG_TYPE"},"nameZh":"药物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062221855140,"alterId":2025062221855140}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Medicine","identityType":"SPG_TYPE"},"nameZh":"药物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062221855140,"alterId":2025062221855140}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Medicine","identityType":"SPG_TYPE"},"nameZh":"药物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062221855140,"alterId":2025062221855140}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062253268986,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062215934835,"alterId":2025062215934835},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Medicine","identityType":"SPG_TYPE"},"nameZh":"药物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062221855140,"alterId":2025062221855140}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062297493280,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062256836182,"alterId":2025062256836182},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062221855140,"alterId":2025062221855140},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期"},"parentTypeInfo":{"uniqueId":2025062281314318,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062281314318]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062281314318,"alterId":2025062281314318}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062281314318,"alterId":2025062281314318}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062281314318,"alterId":2025062281314318}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062281314318,"alterId":2025062281314318}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062218822049,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062212503079,"alterId":2025062212503079},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062281314318,"alterId":2025062281314318}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062215533400,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062237346563,"alterId":2025062237346563},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062281314318,"alterId":2025062281314318},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Keyword","identityType":"SPG_TYPE"},"nameZh":"关键词"},"parentTypeInfo":{"uniqueId":2025062271248850,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062271248850]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Keyword","identityType":"SPG_TYPE"},"nameZh":"关键词"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062271248850,"alterId":2025062271248850}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Keyword","identityType":"SPG_TYPE"},"nameZh":"关键词"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062271248850,"alterId":2025062271248850}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Keyword","identityType":"SPG_TYPE"},"nameZh":"关键词"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062271248850,"alterId":2025062271248850}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Keyword","identityType":"SPG_TYPE"},"nameZh":"关键词"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062271248850,"alterId":2025062271248850}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062283916582,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062249293518,"alterId":2025062249293518},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Keyword","identityType":"SPG_TYPE"},"nameZh":"关键词"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062271248850,"alterId":2025062271248850}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062238474191,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062290439227,"alterId":2025062290439227},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062271248850,"alterId":2025062271248850},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"ArtificialObject","identityType":"SPG_TYPE"},"nameZh":"人造物体"},"parentTypeInfo":{"uniqueId":2025062233367856,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062233367856]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"ArtificialObject","identityType":"SPG_TYPE"},"nameZh":"人造物体"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062233367856,"alterId":2025062233367856}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"ArtificialObject","identityType":"SPG_TYPE"},"nameZh":"人造物体"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062233367856,"alterId":2025062233367856}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"ArtificialObject","identityType":"SPG_TYPE"},"nameZh":"人造物体"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062233367856,"alterId":2025062233367856}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"ArtificialObject","identityType":"SPG_TYPE"},"nameZh":"人造物体"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062233367856,"alterId":2025062233367856}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062240702279,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062249965316,"alterId":2025062249965316},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"ArtificialObject","identityType":"SPG_TYPE"},"nameZh":"人造物体"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062233367856,"alterId":2025062233367856}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062258049744,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062296294459,"alterId":2025062296294459},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062233367856,"alterId":2025062233367856},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Organization","identityType":"SPG_TYPE"},"nameZh":"组织机构"},"parentTypeInfo":{"uniqueId":2025062250740582,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062250740582]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Organization","identityType":"SPG_TYPE"},"nameZh":"组织机构"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250740582,"alterId":2025062250740582}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Organization","identityType":"SPG_TYPE"},"nameZh":"组织机构"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250740582,"alterId":2025062250740582}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Organization","identityType":"SPG_TYPE"},"nameZh":"组织机构"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250740582,"alterId":2025062250740582}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Organization","identityType":"SPG_TYPE"},"nameZh":"组织机构"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250740582,"alterId":2025062250740582}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062228071277,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062289580715,"alterId":2025062289580715},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Organization","identityType":"SPG_TYPE"},"nameZh":"组织机构"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250740582,"alterId":2025062250740582}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062214323476,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062296221722,"alterId":2025062296221722},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062250740582,"alterId":2025062250740582},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Astronomy","identityType":"SPG_TYPE"},"nameZh":"天文学"},"parentTypeInfo":{"uniqueId":2025062214088562,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062214088562]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Astronomy","identityType":"SPG_TYPE"},"nameZh":"天文学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214088562,"alterId":2025062214088562}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Astronomy","identityType":"SPG_TYPE"},"nameZh":"天文学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214088562,"alterId":2025062214088562}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Astronomy","identityType":"SPG_TYPE"},"nameZh":"天文学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214088562,"alterId":2025062214088562}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Astronomy","identityType":"SPG_TYPE"},"nameZh":"天文学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214088562,"alterId":2025062214088562}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062235815737,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062222277700,"alterId":2025062222277700},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Astronomy","identityType":"SPG_TYPE"},"nameZh":"天文学"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214088562,"alterId":2025062214088562}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062231946497,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062285896448,"alterId":2025062285896448},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062214088562,"alterId":2025062214088562},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Concept","identityType":"SPG_TYPE"},"nameZh":"概念"},"parentTypeInfo":{"uniqueId":2025062250721389,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062250721389]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Concept","identityType":"SPG_TYPE"},"nameZh":"概念"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250721389,"alterId":2025062250721389}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Concept","identityType":"SPG_TYPE"},"nameZh":"概念"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250721389,"alterId":2025062250721389}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Concept","identityType":"SPG_TYPE"},"nameZh":"概念"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250721389,"alterId":2025062250721389}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Concept","identityType":"SPG_TYPE"},"nameZh":"概念"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250721389,"alterId":2025062250721389}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062287604350,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062279341758,"alterId":2025062279341758},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Concept","identityType":"SPG_TYPE"},"nameZh":"概念"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062250721389,"alterId":2025062250721389}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062260006436,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062284823597,"alterId":2025062284823597},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062250721389,"alterId":2025062250721389},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"parentTypeInfo":{"uniqueId":2025062235203130,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062235203130]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062235418188,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062235955348,"alterId":2025062235955348},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062272371393,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062297075921,"alterId":2025062297075921},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130},"extInfo":{"commonIndex":false}},{"@type":"EVENT_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"parentTypeInfo":{"uniqueId":2025062247533437,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062247533437]},"spgTypeEnum":"EVENT_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"type","identityType":"PREDICATE"},"nameZh":"事件类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062255337020,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062217109313,"alterId":2025062217109313},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"subject","identityType":"PREDICATE"},"nameZh":"主体"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","propertyGroup":"SUBJECT","constraint":{"id":2025062267527884,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[]},"projectId":1,"ontologyId":{"uniqueId":2025062233749278,"alterId":2025062233749278},"extInfo":{"valueType":"ENTITY_TYPE","propertyTag":"SUBJECT"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"abstract","identityType":"PREDICATE"},"nameZh":"摘要"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062212101157,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062246988950,"alterId":2025062246988950},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"participants","identityType":"PREDICATE"},"nameZh":"参与者"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062252101765,"constraintItems":[{"@type":"MULTI_VALUE"}]},"withIndex":false,"subProperties":[],"semantics":[]},"projectId":1,"ontologyId":{"uniqueId":2025062249777670,"alterId":2025062249777670},"extInfo":{"valueType":"ENTITY_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"location","identityType":"PREDICATE"},"nameZh":"地点"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"GeographicLocation","identityType":"SPG_TYPE"},"nameZh":"地理位置"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062218643714,"alterId":2025062218643714}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062214192646,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[]},"projectId":1,"ontologyId":{"uniqueId":2025062257129536,"alterId":2025062257129536},"extInfo":{"valueType":"ENTITY_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"eventTime","identityType":"PREDICATE"},"nameZh":"发生时间","desc":"发生时间"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"STD","nameEn":"Timestamp","identityType":"SPG_TYPE"},"nameZh":"时间戳","desc":"10位或者13位的时间戳"},"spgTypeEnum":"STANDARD_TYPE","ontologyId":{"uniqueId":29,"alterId":29}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","propertyGroup":"TIME","withIndex":false,"subProperties":[],"semantics":[]},"projectId":1,"ontologyId":{"uniqueId":2025062267232298,"alterId":2025062267232298},"extInfo":{"valueType":"STANDARD_TYPE","propertyTag":"TIME"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"time","identityType":"PREDICATE"},"nameZh":"时间"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062281314318,"alterId":2025062281314318}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062286074494,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[]},"projectId":1,"ontologyId":{"uniqueId":2025062296496885,"alterId":2025062296496885},"extInfo":{"valueType":"ENTITY_TYPE"}}],"relations":[{"isDynamic":true,"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"subject","identityType":"PREDICATE"},"nameZh":"主体"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","propertyGroup":"SUBJECT","constraint":{"id":2025062267527884,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[]},"projectId":1,"ontologyId":{"uniqueId":2025062233749278,"alterId":2025062233749278},"extInfo":{"valueType":"ENTITY_TYPE","propertyTag":"SUBJECT"}},{"isDynamic":true,"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"participants","identityType":"PREDICATE"},"nameZh":"参与者"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Person","identityType":"SPG_TYPE"},"nameZh":"人物"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062235203130,"alterId":2025062235203130}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062252101765,"constraintItems":[{"@type":"MULTI_VALUE"}]},"withIndex":false,"subProperties":[],"semantics":[]},"projectId":1,"ontologyId":{"uniqueId":2025062249777670,"alterId":2025062249777670},"extInfo":{"valueType":"ENTITY_TYPE"}},{"isDynamic":true,"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"location","identityType":"PREDICATE"},"nameZh":"地点"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"GeographicLocation","identityType":"SPG_TYPE"},"nameZh":"地理位置"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062218643714,"alterId":2025062218643714}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062214192646,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[]},"projectId":1,"ontologyId":{"uniqueId":2025062257129536,"alterId":2025062257129536},"extInfo":{"valueType":"ENTITY_TYPE"}},{"isDynamic":true,"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Event","identityType":"SPG_TYPE"},"nameZh":"事件"},"spgTypeEnum":"EVENT_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437}},"basicInfo":{"name":{"@type":"PREDICATE","name":"time","identityType":"PREDICATE"},"nameZh":"时间"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Date","identityType":"SPG_TYPE"},"nameZh":"日期"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062281314318,"alterId":2025062281314318}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062286074494,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[]},"projectId":1,"ontologyId":{"uniqueId":2025062296496885,"alterId":2025062296496885},"extInfo":{"valueType":"ENTITY_TYPE"}}],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062247533437,"alterId":2025062247533437},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"GeographicLocation","identityType":"SPG_TYPE"},"nameZh":"地理位置"},"parentTypeInfo":{"uniqueId":2025062218643714,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062218643714]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"GeographicLocation","identityType":"SPG_TYPE"},"nameZh":"地理位置"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062218643714,"alterId":2025062218643714}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"GeographicLocation","identityType":"SPG_TYPE"},"nameZh":"地理位置"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062218643714,"alterId":2025062218643714}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"GeographicLocation","identityType":"SPG_TYPE"},"nameZh":"地理位置"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062218643714,"alterId":2025062218643714}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"GeographicLocation","identityType":"SPG_TYPE"},"nameZh":"地理位置"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062218643714,"alterId":2025062218643714}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062218250984,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062227351731,"alterId":2025062227351731},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"GeographicLocation","identityType":"SPG_TYPE"},"nameZh":"地理位置"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062218643714,"alterId":2025062218643714}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062282378104,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062261351864,"alterId":2025062261351864},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062218643714,"alterId":2025062218643714},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Others","identityType":"SPG_TYPE"},"nameZh":"其他"},"parentTypeInfo":{"uniqueId":2025062265636877,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062265636877]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Others","identityType":"SPG_TYPE"},"nameZh":"其他"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062265636877,"alterId":2025062265636877}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Others","identityType":"SPG_TYPE"},"nameZh":"其他"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062265636877,"alterId":2025062265636877}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Others","identityType":"SPG_TYPE"},"nameZh":"其他"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062265636877,"alterId":2025062265636877}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Others","identityType":"SPG_TYPE"},"nameZh":"其他"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062265636877,"alterId":2025062265636877}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062272928037,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062212532553,"alterId":2025062212532553},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Others","identityType":"SPG_TYPE"},"nameZh":"其他"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062265636877,"alterId":2025062265636877}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062229637492,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062215594172,"alterId":2025062215594172},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062265636877,"alterId":2025062265636877},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Transport","identityType":"SPG_TYPE"},"nameZh":"运输"},"parentTypeInfo":{"uniqueId":2025062214582338,"parentUniqueId":1,"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"},"inheritPath":[1,2025062214582338]},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Transport","identityType":"SPG_TYPE"},"nameZh":"运输"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214582338,"alterId":2025062214582338}},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":1,"alterId":1},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Transport","identityType":"SPG_TYPE"},"nameZh":"运输"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214582338,"alterId":2025062214582338}},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":2,"alterId":2},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Transport","identityType":"SPG_TYPE"},"nameZh":"运输"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214582338,"alterId":2025062214582338}},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":true,"advancedConfig":{"encryptTypeEnum":"NONE","withIndex":false,"subProperties":[],"semantics":[]},"projectId":0,"ontologyId":{"uniqueId":3,"alterId":3},"extInfo":{}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Transport","identityType":"SPG_TYPE"},"nameZh":"运输"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214582338,"alterId":2025062214582338}},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062266164975,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT_AND_VECTOR"},"projectId":1,"ontologyId":{"uniqueId":2025062253872120,"alterId":2025062253872120},"extInfo":{"valueType":"BASIC_TYPE"}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","namespace":"TwoWikiTest","nameEn":"Transport","identityType":"SPG_TYPE"},"nameZh":"运输"},"spgTypeEnum":"ENTITY_TYPE","projectId":1,"ontologyId":{"uniqueId":2025062214582338,"alterId":2025062214582338}},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE","ontologyId":{"uniqueId":2,"alterId":2}},"inherited":false,"advancedConfig":{"encryptTypeEnum":"NONE","constraint":{"id":2025062224900454,"constraintItems":[]},"withIndex":false,"subProperties":[],"semantics":[],"indexType":"TEXT"},"projectId":1,"ontologyId":{"uniqueId":2025062297762118,"alterId":2025062297762118},"extInfo":{"valueType":"BASIC_TYPE"}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"},"projectId":1,"ontologyId":{"uniqueId":2025062214582338,"alterId":2025062214582338},"extInfo":{"commonIndex":false}},{"@type":"ENTITY_TYPE","basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Entity","identityType":"SPG_TYPE"}},"parentTypeInfo":{"parentTypeIdentifier":{"@type":"SPG_TYPE","nameEn":"Thing","identityType":"SPG_TYPE"}},"spgTypeEnum":"ENTITY_TYPE","properties":[{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Entity","identityType":"SPG_TYPE"}},"spgTypeEnum":"ENTITY_TYPE"},"basicInfo":{"name":{"@type":"PREDICATE","name":"description","identityType":"PREDICATE"},"nameZh":"描述","desc":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE"},"inherited":false,"advancedConfig":{"withIndex":false,"subProperties":[],"semantics":[]}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Entity","identityType":"SPG_TYPE"}},"spgTypeEnum":"ENTITY_TYPE"},"basicInfo":{"name":{"@type":"PREDICATE","name":"id","identityType":"PREDICATE"},"nameZh":"实体主键","desc":"实体主键"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE"},"inherited":false,"advancedConfig":{"withIndex":false,"subProperties":[],"semantics":[]}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Entity","identityType":"SPG_TYPE"}},"spgTypeEnum":"ENTITY_TYPE"},"basicInfo":{"name":{"@type":"PREDICATE","name":"name","identityType":"PREDICATE"},"nameZh":"名称","desc":"名称"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE"},"inherited":false,"advancedConfig":{"withIndex":false,"subProperties":[],"semantics":[]}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Entity","identityType":"SPG_TYPE"}},"spgTypeEnum":"ENTITY_TYPE"},"basicInfo":{"name":{"@type":"PREDICATE","name":"desc","identityType":"PREDICATE"},"nameZh":"描述","desc":"描述"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE"},"inherited":false,"advancedConfig":{"withIndex":false,"subProperties":[],"semantics":[]}},{"subjectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Entity","identityType":"SPG_TYPE"}},"spgTypeEnum":"ENTITY_TYPE"},"basicInfo":{"name":{"@type":"PREDICATE","name":"semanticType","identityType":"PREDICATE"},"nameZh":"语义类型","desc":"语义类型"},"objectTypeRef":{"basicInfo":{"name":{"@type":"SPG_TYPE","nameEn":"Text","identityType":"SPG_TYPE"},"nameZh":"文本","desc":"文本"},"spgTypeEnum":"BASIC_TYPE"},"inherited":false,"advancedConfig":{"withIndex":false,"subProperties":[],"semantics":[]}}],"relations":[],"advancedConfig":{"visibleScope":"DOMAIN"}}]}
