+++ knext/schema/client.py
@@ -11,13 +11,15 @@
 # or implied.
 import os
 from typing import List, Dict
-
+import re
 import knext.common.cache
 from knext.common.base.client import Client
 from knext.common.rest import ApiClient, Configuration
 from knext.schema import rest
 from knext.schema.model.base import BaseSpgType, AlterOperationEnum, SpgTypeEnum
 from knext.schema.model.relation import Relation
+
+from kag.jiuyuansolver.jygraph_apiclient import ProjectSchemaClient
 
 import json
 
@@ -51,9 +53,19 @@ class SchemaSession:
         #     self._project_id
         # )
 
-        with open("/root/softwares/kag_project/KAG-master/kag/jiuyuansolver/schema3.py", 'r', encoding='utf-8') as file:
-            content = file.read()
-        project_schema = ApiClient().jiuyuan_deserialize(content, "ProjectSchema")
+        BASE_URL = 'http://172.22.162.15:8090'  # 替换为实际API域名
+        client = ProjectSchemaClient(BASE_URL)
+        response_data = None
+        try:
+            project_id = '1' 
+            result = client.query_project_schema(project_id)
+            response_data = result.decode("utf-8")
+        except Exception as e:
+            print(f"查询失败: {e}")    
+
+        # with open("/root/softwares/kag_project/KAG-master/kag/jiuyuansolver/schema2.py", 'r', encoding='utf-8') as file:
+        #     content = file.read()
+        project_schema = ApiClient().jiuyuan_deserialize(response_data, "ProjectSchema")
 
         for spg_type in project_schema.spg_types:
             spg_type_name = spg_type.basic_info.name.name
@@ -68,6 +80,12 @@ class SchemaSession:
                 self._spg_types[spg_type_name] = type_class(
                     name=spg_type_name, rest_model=spg_type
                 )
+
+        # print("self._spg_types")
+        # print(self._spg_types)
+
+        # print("self.__spg_types")
+        # print(self.__spg_types)
 
     @property
     def spg_types(self) -> Dict[str, BaseSpgType]:
