+++ knext/schema/rest/schema_api.py
@@ -258,6 +258,8 @@ class SchemaApi(object):
 
         header_params = {}
 
+        print(query_params)
+
         form_params = []
         local_var_files = {}
